feat: Implement memory management functions including malloc and free
